requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
pypdf2==3.0.1
pymongo==4.6.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0
zipfile36==0.1.3
numpy==1.26.4
pandas==2.1.1
